Move serializeString and parseString to SerializationUtils in sdk.utils
